requirements.txt
scons
conan
pre-commit
